Adds support for folder/* patterns in .fleetignore
As of now, the `folder/*` was not supported by Fleet.

This PR adds support for it, giving the option to ignore folders (not files with a given name).

For example:

This folder structure:

```
root
   dir1
     test
       file1
       file2
   file3
test
file4
```

with the following `.fleetignore` file:
```
test/*
```

would produce the following recource list:
```
root/file3
root/test
root/file4
```

`test` folder would be ignored but `test` file would not.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>