Use RequeueAfter for Gitpolling job.
Uses the Manager's timers instead of creating a different go routine to deal with the polling job.

This is an approach to deal with multiple calls updating the status of a `gitrepo` in the same reconcile loop.
There were issues setting the status and getting it in the same reconcile as the cache was not properly set yet.

This tries to fix those issues by only updating the status once per reconcile.

Refers to: https://github.com/rancher/fleet/issues/2631

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>