Extract git job node selector from deployment
This prevents the gitOps controller from needing to query running pods
to extract a possible node selector from the definition of a `gitjob` or
`fleetcontroller` pod running for its shard.

Instead, the `gitjob` deployment feeds the sharding-related node
selector, if any, to `fleetcontroller gitjob` as a JSON-formatted
command-line argument, from where the gitOps controller can unmarshal
it and forward it to created jobs.