Skip content check for bundledeployment deletion
When directly deleting a bundle deployment resource, which has been
created from a `GitRepo` through a bundle, the bundle controller
re-creates the content resource when that content resource is deleted,
because a bundle still exists.

This means that checking for content resource deletion depends on a race
condition, ie on whether test logic can validate that the resource has
been deleted before Fleet re-creates it. Bypassing this would require
directly creating a bundle deployment on its own which, on the other
hand, is not a realistic Fleet use case. Therefore, this specific check
for existence of a content resource is skipped for that test case, while
it is kept in cases where the content resource is expected to be deleted
because the parent `GitRepo` or bundle has been deleted.