Clean up content resources through finalizers (#2566)
Content resources may be used by one or more bundle deployments, which
now each add their own finalizer on a content resource to indicate their
reliance on it, and delete that finalizer when that reliance no longer
applies.

This eliminates the need for batch cleanup of content resources.

Content deletion is now tested in a test suite related to finalizers,
which somewhat simplifies checks around regular `GitRepo` test suites
and prevents Content deletion tests from requiring test infrastructure.

This uncovered a bug, fixed here, whereby direct bundle deployment
deletion would not delete the corresponding Content resource.

When directly deleting a bundle deployment resource, which has been
created from a `GitRepo` through a bundle, the bundle controller
re-creates the content resource when that content resource is deleted,
because a bundle still exists.

This means that checking for content resource deletion depends on a race
condition, ie on whether test logic can validate that the resource has
been deleted before Fleet re-creates it. Bypassing this would require
directly creating a bundle deployment on its own which, on the other
hand, is not a realistic Fleet use case. Therefore, this specific check
for existence of a content resource is skipped for that test case, while
it is kept in cases where the content resource is expected to be deleted
because the parent `GitRepo` or bundle has been deleted.