[v0.10] Fixes finalizer deletion in bundle deployments
This issue is affecting the `BundleDeployment` when it is deleted after targetting discards it as a valid target.

The reconciler was not called when `DeletionTimestamp` was set and there was a typo that prevented the finalizer to be deleted.

Refers to: https://github.com/rancher/fleet/issues/2768

Backport of: https://github.com/rancher/fleet/pull/2820

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>