Fix helm.sh/resource-policy being added to everything and not just CRDs
When `deleteCRDResources` was disabled (default), the
`helm.sh/resource-policy` annotation was added to almost all resources
instead of just to the CRDs as it was intended. The fix is to not change
the shared `annotations` that also get added to other resources, but
only modify a copy that's only used for one resource.

Fixes #2716