Add integration test checks for CA bundle secret (#2846)
When a git job is created for a `GitRepo` with a configured CA bundle,
integration tests now validate that a secret is created with the
contents of that bundle and validate its contents against the `GitRepo`
spec.

This also validates through integration tests that RBAC resources and
the CA bundle secret are owned by their `GitRepo` and will therefore
be deleted as soon as the `GitRepo` is.
Validation does not actually delete anything, due to a limitation [1] of
the test framework.

[1]: https://book.kubebuilder.io/reference/envtest#testing-considerations