Deletes retries in the gitops job.
This PR deletes retries in the gitops job created to call `gitCloner` + `fleet apply`.
The user experience is affected because we need to wait for all the retries to finish before we can show up any possible error.

Jobs created are managed by the job controller in kubernetes, so we don't have control on the retry pods or when the job is finally considered as failed.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>