Convert drift e2e tests into integration tests (#2861)
* Simplify drift correction integration tests

This removes checks on bundle statuses, which are unrelated to drift
correction and already covered in status tests
(`integrationtests/agent/bundle_deployment_status_test.go`).

* Migrate drift correction end-to-end tests

This migrates end-to-end test cases for drift correction to
integration tests, simplifying them and reducing the risk of flakiness.

* Rename `createBundle` as `createBundleDeployment`

That bundle reconciler method creates bundle deployments, not bundles.

* Convert bundle deletion e2e test into integration test

That end-to-end test case was flaky, and could easily be migrated to
integration tests.