Test Fleet in Rancher on self-hosted runner (#2804)
* Create new workflow for testing Fleet in Rancher

This leaves the original workflow in place, upgrading Fleet in Rancher.
The new workflow runs on Rancher's org-wide, non-containerised
hosted runners. It installs the latest Rancher stable (non-alpha) version,
pointing it to a configurable version of Fleet (which can be built eg. from
the workflow for building test Fleet charts from any commit).

* Install Rancher via Helm

This simplifies Rancher installation, preventing a costly local build
and simply making use of existing configuration options to override
the Fleet version to install from a custom repository and branch.

The Rancher Docker image to use is hard-coded for now.

* Grant cluster-admin privilege to downstream cluster admin

This enables downstream cluster registration to succeed, as specified in
the official Rancher docs [1].

[1]: https://ranchermanager.docs.rancher.com/v2.0-v2.4/how-to-guides/new-user-guides/kubernetes-clusters-in-rancher-setup/import-existing-clusters#prerequisites

* Require a non-empty command to import a cluster

The Rancher CLI may output an empty command for a bit, leading
`register-downstream-clusters.sh` to fail. Instead of trying to reverse
engineer why that might be, we simply run `rancher cluster import`
repeatedly until the returned command is non-empty.

* Fix downstream cluster registration

When installing Rancher through Helm:
* the `CATTLE_SERVER_URL` needs to be set to the same value as `hostname`
* TLS mode must be set to `system-store`, to prevent cert-related errors
  when running a Fleet agent in a downstream cluster

* Ensure API server IP is available before setting it

This waits for the upstream cluster to be ready, preventing an empty IP
from being set.