Adds predicate when webhook commit changes
This adds a new predicate when the webhook commit is changed in the `GitRepo`.
It also filters out `Job` events for creation/deletion, because they add extra reconcile loops
that only increase the possibility of race conditions and that are not required.

The reconciler already knows when a `Job` is created/deleted because it is the owner and does not
need to react upon those 2 events from the cluster.

Refers to: https://github.com/rancher/fleet/issues/2969

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>