Schedule workflow releasing test charts
Scheduling the workflow running end-to-end tests against Rancher would
not work, as Rancher would be installed with an empty Fleet version
setting, not find that Fleet version, and time out.

Instead, this commit schedules the workflow releasing test charts for
the head of `rancher/fleet`'s `main` branch, which then calls the test
workflow with the right version, repo and branch parameters for Rancher
to install those test charts.