GitRepo resources list doesn't list resources multiple times (#3027)
GitRepo resources list doesn't list resources multiple times

This is a hotfix to avoid a conflict with UI code, which multiplies the resources list per cluster. Multiplying a GitRepo's resources from all clusters with the number of clusters potentially freezes the UI.

With this PR gitrepo.Status.Resources reports resources only once, even if they are deployed to multiple clusters. If any cluster has a nonready/modified status for the resource, it is still added to the PerClusterState field.

That means, it's no longer possible to tell if a resource was not deployed to a cluster, e.g. because it was excluded via helm chart values.