Skip install loop for invalid patches
This should prevent the Fleet agent from repeatedly trying to install a
release with an invalid patch when trying to undo drift on a service's
ports, which can only be rolled back in force mode so far, ie by
uninstalling and reinstalling the release.
In turn, this may prevent the original, valid release revision from
disappearing, which caused discrepancies in version numbers between Helm
history and the bundle deployment status, leading to sets of resources
being emptied and eventually to bundles appearing as ready for drifting
bundle deployments.