Inject registered cluster name into multi-cluster tests
Depending on the context in which multi-cluster end-to-end tests are
run, there may not be any registered downstream cluster called `second`.
In such cases, such as when testing Fleet in Rancher, the CI workflow
computes the name of the registered cluster and exports it as an
environment variable, for the tests to use instead of `second`.