Fix base Fleet chart version resolution for test release (#3068)
When releasing test Fleet charts, the test release workflow looks for
the latest existing Fleet release, to use it as a base before making a
few edits.

The previous logic used to find the latest available chart was buggy, in
that it would list releases in alphabetical order, which could differ
from semver. For instance, chart version `103.1.10+up0.9.11` would be
listed between versions `103.1.0+up0.9.0` and `103.1.2+up0.9.2`.

Instead, this commit simplifies resolution by first looking at the
`package.yaml` file, extracting the chart version from there and looking
for the corresponding Fleet version in the charts repository.
Resolution would then fail if no corresponding version is found in the
repository, but that is far less likely to happen than with the previous
logic and would typically be a symptom of a broken state of the charts
repository.