Run node selector check earlier
Sharding end-to-end tests exhibited flakiness, caused by the git job pod
not being present in the cluster by the time checks were run to validate
its node selector against that of the relevant shard.

To prevent this, tests on the node selector are now run first, which
incidentally also prevents `Eventually` from running more times awaiting
a config map to be deployed.