Fix and document the code
The fix is about situations in which a Bundle has Bundle Deployments,
i.e. the template values could correctly been rendered. But then the
a cluster label is changed that is used in templating, which produces an
error. Now there is an error on the Bundle that is supposed to be copied
to the GitRepo (which, when I now think about it, doesn't work, since
the GitRepo can have more than one Bundle! It would also need to
properly aggregate those errors from all it's Bundles, possibilty
limitting the amount to 10 errors, like in other places in the code)
which can be different from the errors in the Bundle Deployments (if
they have any).