Enables e2e tests using k3d services export to external ips
This PR changes some e2e environment files to be able to run e2e tests
in systems in which we cannot access load balancers IPs from clusters.

This is specially intented to be able to execute the e2e tests locally
in MAC OS systems in which docker IPs cannot be accessed.

* Get your IP address (localhost is not valid. See
  [this](https://github.com/helm/helm/issues/11683#issuecomment-2183304555)
* Export your IP with export external_ip=YOUR_IP or edit the
  dev/env.single-cluster-defaults file and add it.
* Run e2e tests normally

Note: when running tests with external_ip set ports passed to `k3d` are
not random. This in intentional as we need to access the services in a
well known port.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>

tt