Exports Metrics URLs to test with external IPs
It also deletes the HELM_PATH env variable as it is no longer used.

Fixes errors in metrics tests when the test is trying to check for metrics
when the service exporting them it's still not fully up.

Adds a custom file to add extra env variables

Adds an extra check in OCI e2 tests to verify that the `CI_OCI_USERNAME` and
`CI_OCI_PASSWORD` are set.

Deletes the `GI_GIT_REPO_URL` env variable and uses the already existing `external_ip`.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>