Fall back to Rancher CA bundles in git cloner
This makes use of the new `cert` package to use a CA bundle from Rancher
secrets, if configured and if the `GitRepo` does not already specify any
such CA bundle.
This does not alter the git cloner itself, but allows the CA bundle
secret built by the gitOps controller (and later mounted for use by the
git cloner) to use CA bundle information extracted from Rancher secrets
as an alternative source.