Adds new gitops poller controller
Adds a new poller controller to avoid sharing events between the main `GitRepo` controller and the polling mechanism.
New poller controller only does polling and resets its state upon generation change.

When a new commit is found by the poller controller it sets the `PollerCommit` field in the `GitRepo`'s status.
The main controller watches for changes in that field and it is called to reflect the new commit in the `GitRepo`'s `Status.Commit` field.

A new `PollerGeneration` field is added so it does not interfere with the already existing `ObservedGeneration` used in the main `GitRepo` controller

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>