Fall back to Rancher CA bundle with incomplete Helm secret
This caters for an edge case whereby a `GitRepo` may specify a Helm
secret, for instance through `HelmSecretName`, which does not contain a
CA bundle.
When that happens, Fleet will fall back to Rancher-configured CA
bundles, if any, while any existing CA bundle in the `GitRepo`'s Helm
secret would, as before, be used with higher precedence.