[v0.10] - Always returns result with RequeueAfter set
Changes the gitops reconciler result to always have the polling interval in RequeueAfter.
This way we take advantage of the fact that if there is already an event in the RequeueAfter queue the next ones will be dropped and we won't miss the polling cycle if there is a race condition between RequeueAfter events and any other event in the reconciler.

This is only added when `DisablePolling` is not set.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>