Copy urlError in Bundle annotations to status
This leads to the GitRepo controller picking up the status message from
the Bundle, so that the GitRepo resource also shows the same error that
is shown in the Bundle.

This works without having to further adapt the GitRepo but likely is the
same mechanism that copies the template errors from the Bundle to the
GitRepo.