Change ImageScanCommit to a pointer in the GitRepo Spec
Changes ImageScanCommit to a pointer because otherwise default values are stored on the first time a `GitRepo` is updated and generation is changed, which means another call to the reconciler.

This creates a double call to fetch for new commits (poll the `GitRepo`) when a `GitRepo` is added.

On the first time it will poll because `LastPollingTime` in Status is not set yet.
And the second time it will fetch for new commits is when generations changes from 1 to 2 (which is almost immediate after the `GitRepo` is updated to add the finalizer.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>