Change ImageScanCommit in GitRepo spec to a pointer
As ImageScanCommit is an optional field inside the GitRepo spec this PR changes it to be a pointer.

Not doing so means that the veru first time the `GitRepo` is stored in the cluster default values will be created
and generation will be incresed, which means a double polling for the `GitRepo` when it is created.

The first time the `GitRepo` is polled because the `LastPollingTime` is not set.
The second time it is polled because of the generation change mentioned above.

Aside from double polling in a short period of time this PR also tries to prevent an extra reconciler call that is not needed.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>