Adds tolertions to GitRepo
When applying a `GitRepo` in a cluster where all the nodes are tainted the pod created to call `fleet apply` remans waiting as Pending.

This PR adds tolerations to the `GitRepo` spec. Those tolerations will be added to the job spec when running `fleet apply` so the pod is scheduled.

Refers to: https://github.com/rancher/fleet/issues/3313

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>