Fall back to Rancher-configured CA bundles (#3233)
This enables Fleet to rely on Rancher-configured CA bundle secrets [1]
in the following clients:
* git remote lister and cloner: those CA bundles no longer need to be
  provided in a `GitRepo`'s spec. Fleet will extract CA bundle data from
  Rancher secrets if (and only if) no CA bundle is provided for a
  `GitRepo`.
* Helm client: Fleet will extract, populate and mount Rancher CA bundles
  in the following cases:
    * no `HelmSecretName` nor `HelmSecretNameForPaths` secret is
      provided on a `GitRepo`
    * one of those secrets is provided, but only specifies username(s)
      and password(s), without a CA bundle

## Limitations

This does not cover imagescan (pending further discussion on the future
of that feature), nor the HelmOps controller.

## Squashed commits

* Add cert package to extract Rancher CA bundle info

This commit provides a new `cert` package, containing logic for
extraction of CA bundle information from known Rancher secrets.
That new package is expected to be used in Fleet clients which will need
to fall back to Rancher CA bundles when no custom bundle is provided.

* Fall back to Rancher CA bundles in git commit fetcher

This makes use of the new `cert` package to use a CA bundle from Rancher
secrets, if configured and if the `GitRepo` does not specify any such CA
bundle.

* Fall back to Rancher CA bundles in git cloner

This makes use of the new `cert` package to use a CA bundle from Rancher
secrets, if configured and if the `GitRepo` does not already specify any
such CA bundle.
This does not alter the git cloner itself, but allows the CA bundle
secret built by the gitOps controller (and later mounted for use by the
git cloner) to use CA bundle information extracted from Rancher secrets
as an alternative source.

* Fall back to Rancher CA bundles for Helm client

This makes use of the new `cert` package to use a CA bundle from Rancher
secrets, if configured and if the `GitRepo` does not already specify any
Helm secret. If those conditions are met, the gitOps controller creates
a new secret with CA bundle information from Rancher.

This does not alter the Helm client itself, and could be optimised to
avoid duplicating the CA bundle secret in case it already exists, with
identical contents, due to no CA bundle being configured on the
`GitRepo`.

* Mount Rancher CA bundle data for git cloner

This had been omitted in a previous commit, leading to the git cloner
being unable to clone a git repository hosted on a server using HTTPS
with certificates signed by a private CA.

* Add end-to-end test for falling back to Rancher CA bundles

This covers fallback in git clients. It uses an updated version of the
test git server, which serves HTTPS, beside HTTP, with a custom CA
bundle.

* Fall back to Rancher CA bundle with incomplete Helm secret

This caters for an edge case whereby a `GitRepo` may specify a Helm
secret, for instance through `HelmSecretName`, which does not contain a
CA bundle.
When that happens, Fleet will fall back to Rancher-configured CA
bundles, if any, while any existing CA bundle in the `GitRepo`'s Helm
secret would, as before, be used with higher precedence.

* Automate creation of test Rancher secret

This makes use of new Rancher-like secret [1] creation logic when
setting up Fleet for end-to-end testing purposes. This allows tests
validating fallback to Rancher-configured CA bundles to pass.

[1]: https://ranchermanager.docs.rancher.com/getting-started/installation-and-upgrade/installation-references/helm-chart-options#additional-trusted-cas