Add Lifecycle secrets
Each Bundle and BundleDeployment can have a secret to store options in.
The secrets have the same name and namespace as the resource.
For now they only store the helm values.

* Remove "fleet apply" config files from bundle resources (#3403)

This will make the bundle and content resources smaller. However, it
should not change the helm chart that is deployed. The files are parsed
into deployment options and have no use after the fleet apply stage.

Minor changes:
* Extract loaddirectory args to struct
* Remove unused key
* Clean up creation of agent manifests
* Separate bundledeployment construction from creation
* Remove slow unit test
  This test relied on a network timeout and took up to 3 minutes.