Slightly speed up end-to-end test runs  (#3443)
* Reduce setup loops in end-to-end tests

This uses `By` statements to distinguish use cases with common setup,
instead of having each of them live in its own `It` block, which would
lead to that setup logic being run again.

* Add `--reuse` option to dev/setup-single-cluster

This enables that script to reuse a cluster when called, which is faster
than deleting the cluster and re-creating it.