Fix CA bundle mount after GitRepo update
This fixes a bug whereby, once a `GitRepo` had been created, subsequent
reconciles would not detect a Rancher secret as the source of CA bundle
data. Relevant information for the controller is whether a secret
created in the `GitRepo`'s namespace, duplicating a Rancher CA bundle
secret, _exists_, not whether it was created or updated.