Simplify finalizers end-to-end tests
Testing direct bundle deployment deletion does not bring much added
value, as users are not expected to directly deal with that resource.
Besides, content resource deletion is already covered by test cases
deleting a `GitRepo` and a bundle.

To reduce the risk of intermittent test failures, the test driver now
only scales the deployment relevant to the resource being deleted:
`gitjob` for `GitRepo`s and `fleet-controller` for bundles.