Reduce e2e flakiness (#3475)
* Check OCI storage env var on restarted pods

In order to reduce the likelihood of test failures, end-to-end tests now
check that, when amending Fleet deployments to update the experimental
OCI storage environment variable, restarted pods bear the expected
value for that variable.

* Simplify finalizers end-to-end tests

Testing direct bundle deployment deletion does not bring much added
value, as users are not expected to directly deal with that resource.
Besides, content resource deletion is already covered by test cases
deleting a `GitRepo` and a bundle.

To reduce the risk of intermittent test failures, the test driver now
only scales the deployment relevant to the resource being deleted:
`gitjob` for `GitRepo`s and `fleet-controller` for bundles.