New webhook secret handling
This PR changes the secret handling for webhooks in order to add the possibility to define a secret per `GitRepo`.
The actual `main` code has a single global secret shared by all the possible `GitRepos` in the system, which forces the users to share the same secret for all their webhooks in the cluster.

The secret is also cached, which means that it listens to the informer cache for all secrets in the cluster and, if any secret is updated it checks if it's the global secret for webhooks. In that case it reinstantiates all the webhook parsers for all the supported providers with the new secret value (or none if the secret was deleted).

It also needs to instantiate the webhook parsers for all supported providers at the beginning.

Also... a bug was found in that secret caching. If all the `gitjob` pods are restarted, the secret will be completely ignored, accepting any webhook received (which might mean a security problem).

The new implementation does not cache any secret and does not create all the different providers in advance.
It just checks for the secret right before updating the `WebhookCommit` and, depending on the payload, it will instantiate the needed parser for that provider.

The price we pay for this is that we need to double parse the webhook request.
This is because when we receive the payload we don't know yet the name of the `GitRepo` it is trying to update and we cannot check if it has a secret defined for it.

The first round of parsing will be used to determine the name of the `GitRepo` and, if a secret is defined (global or per `GitRepo`) it will parse again setting that secret.

The library used to parse the webhook payload doesn't offer the possibility to check just the secret given a payload.

The new field added in the `GitRepo` CRD is called `webhookSecret` and references the name of the secret, which should be located at the same namespace as the `GitRepo`

If a `Gitrepo` secret is defined it will take preference over the global secret, that is still supported.

This is an example of a `GitRepo` with the new field.

```yaml
apiVersion: fleet.cattle.io/v1alpha1
kind: GitRepo
metadata:
  name: simple
  namespace: fleet-local
spec:
  repo: "https://github.com/rancher/fleet-examples"
  paths:
  - simple
  disablePolling: true
  webhookSecret: webhook-secret-name
```

Refers to: https://github.com/rancher/fleet/issues/2776

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>