Delete no longer targeted bundle deployments
Delete Bundle deployments that exist in the Bundle, that are also deployed, but that are no longer targeted.

When changing either the target customizations or targets in the `GitRepo` definition, Fleet does not delete
bundle deployments that are no longer targeted using the new target rules.

Refers to: https://github.com/rancher/fleet/issues/2995

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>