Calls semaphore Acquire inside go routine
Calls semaphone Acquire inside the go route instead to before the go routine.

Placing the Acquire call before the creation of the goroutine means that if any previously launched goroutine fails, it will cancel the shared context used by Acquire.

As a result, since Acquire is called outside the goroutine, it may return an error due to the canceled context—causing us to propagate a generic "context canceled" error instead of the original error that caused the goroutine to fail.

For example, suppose we're trying to download 10 charts, using a semaphore with a limit of 4.
If the first call to `downloadDirectory` fails, the errgroup will cancel the context.
Then, when the next goroutine attempts to call Acquire, it will receive an error due to the canceled context. This causes the function to return a "context canceled" error, rather than the more informative error from errgroup.Wait(), which would be more helpful to the user.

Additionally, this pull request ensures that we avoid downloading duplicate directories.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>