Deletes parallel directory downloading code.
Removed the code that parallelized directory downloads in order to simplify the implementation and provide more descriptive error messages to the user.

Previously, the semaphore's Acquire call was placed outside the goroutine. This caused any failed goroutine to cancel the shared context, leading subsequent Acquire calls to return a generic "context canceled" error. As a result, the user would see that error instead of the original, more informative error that triggered the cancellation.

Additionally, this pull request ensures that we avoid downloading duplicate directories.

Refers to: https://github.com/rancher/fleet/issues/3234

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>