Use secrets for storing OCI registry details
This PR implements the changes described in https://github.com/rancher/fleet/issues/3688.

Secrets are used to store all the details that define a connection to an OCI registry.
Two levels of secrets are defined: one specified by the user (referenced in the ociReferenceSecret field of GitRepo), and a default one that can be created for a given namespace.

Fleet uses the user-defined secret, and if it does not exist, falls back to the default secret.

A specific type in enforced for any OCI reference secret.

If either of the two secrets exists, a deployment is executed, storing the Bundle resources in the referenced OCI registry. Otherwise, etcd is used.

Two new fields are introduced in the OCI secrets to perform read operations from the OCI registry (these are used by the agents).

A SHA verification of the contents of the manifest downloaded from the OCI registry is added. If the calculated SHA does not match the expected one, the content is rejected and an error is reported.

This PR also includes refactoring of the e2e tests to simplify how cluster resources are read, using a controller-runtime client instead of the kubectl wrapper.

Refers to: https://github.com/rancher/fleet/issues/3688
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>