Reorganise Bundledeployment fields (#3673)
* Separate gitOps-specific bundle deployment options

Those options are not relevant for HelmOps bundles. This commit
therefore separates them into a distinct struct to make this more
explicit, while limiting the risk of breaking the API.

* Move GitOps-specific Helm options to embedded struct

This clarifies that values files are not relevant to HelmOps bundles,
since no repository is used in such cases to store those files.